+++ src/components/FormFields.js
@@ -1,87 +1,67 @@
-import React, { Component } from "react";
+import React, { Component, Fragment } from "react";
 
 import classnames from "../classnames";
+import FormError from "./FormError";
 
-class FormField extends Component {
-  state = { error: null, touched: false };
+const buildFormField = type => {
+  class FormField extends Component {
+    state = { touched: false };
 
-  componentDidMount() {
-    this.deriveError();
-  }
+    handleBlur = () => {
+      this.setState({ touched: true });
+    };
 
-  componentDidUpdate(prevProps) {
-    const updateRequired = ["required", "validator", "value"].some(propName => {
-      const { [propName]: propValue } = this.props;
+    handleChange = ({ target: { value } }) => {
+      const { name, onChange, onFormChange } = this.props;
 
-      return propValue !== prevProps[propName];
-    });
+      if (onChange) {
+        onChange(value);
+      }
 
-    if (updateRequired) {
-      this.deriveError();
+      if (onFormChange) {
+        onFormChange(name, value);
+      }
+    };
+
+    render() {
+      const { name, value } = this.props;
+      const {
+        addon, children, className, onError, onFormChange, required, submitted,
+        validator, ...props
+      } = this.props;
+
+      const { touched } = this.state;
+
+      return (
+        <label className={classnames("chq-ffd", className)} htmlFor={name}>
+          <span className="chq-ffd--lb">{children}</span>
+          <Fragment>
+            {addon && <span className="chq-ffd--ad">{addon}</span>}
+            <input
+              {...props}
+              type={type}
+              id={name}
+              value={value || ""}
+              onBlur={this.handleBlur}
+              onChange={this.handleChange}
+            />
+          </Fragment>
+          <FormError
+            name={name}
+            onError={onError}
+            required={required}
+            submitted={submitted}
+            touched={touched}
+            validator={validator}
+            value={value}
+          />
+        </label>
+      );
     }
   }
 
-  handleBlur = () => {
-    this.setState({ touched: true });
-  };
-
-  handleChange = ({ target: { value } }) => {
-    const { name, onChange, onFormChange } = this.props;
-
-    if (onChange) {
-      onChange(value);
-    }
-
-    if (onFormChange) {
-      onFormChange(name, value);
-    }
-  };
-
-  deriveError = () => {
-    const { name, onError, required, validator, value } = this.props;
-
-    let error = null;
-
-    if (required && !value) {
-      error = "Required";
-    } else if (validator) {
-      error = validator(value);
-    }
-
-    this.setState({ error });
-
-    if (onError) {
-      onError(name, error);
-    }
-  };
-
-  render() {
-    const {
-      addon, children, className, onError, onChange, onFormChange, name,
-      required, submitted, validator, value, ...props
-    } = this.props;
-
-    const { error, touched } = this.state;
-
-    return (
-      <label className={classnames("chq-ffd", className)} htmlFor={name}>
-        <span className="chq-ffd--lb">{children}</span>
-        {addon && <span className="chq-ffd--ad">{addon}</span>}
-        <input
-          {...props}
-          id={name}
-          name={name}
-          value={value || ""}
-          onBlur={this.handleBlur}
-          onChange={this.handleChange}
-        />
-        {error && (submitted || touched) && <p className="chq-ffd--rq">{error}</p>}
-      </label>
-    );
-  }
-}
-
-const buildFormField = type => props => <FormField {...props} type={type} />;
+  return FormField;
+};
 
 export const EmailField = buildFormField("email");
 export const NumberField = buildFormField("number");