+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- The `FileField` component.
+- The `Panel.LoaderBody` subcomponent, for when the body is still being loaded.
+
 ### Changed
 - Increased the `z-index` in the `Nav` component to have it hide the content below it.
 - Improve a11y by adding an `aria-label` to the `Hamburger` button.